+++ src/icon.rs
@@ -1,12 +1,9 @@
-use super::error;
-use super::ztxt;
-use super::RawDmi;
-
 use image::imageops;
 use image::GenericImageView;
 use std::collections::HashMap;
 use std::io::prelude::*;
 use std::io::Cursor;
+use crate::{error, RawDmi, ztxt};
 
 #[derive(Clone, Default)]
 pub struct Icon {
